+++ meetup2xibo/updater/injector.py
@@ -13,8 +13,9 @@ from .place_finder import PlaceFinder
 from .location_chooser import LocationChooser
 from .conflict_analyzer import ConflictAnalyzer, NullConflictAnalyzer
 from .conflict_places import ConflictPlaces, ConflictPlacesLoader
-from .event_converter import EventConverter
+from .event_converter import EventConverter, EventListConverter
 from .event_location import EventLocation
+from .event_suppressor import EventSuppressor
 from .event_updater import EventUpdater
 from .phrase_mapper import PhraseMapper
 from .xibo_api_url_builder import XiboApiUrlBuilder
@@ -146,6 +147,13 @@ def inject_default_event_location(application_scope):
         application_scope.default_places)
 
 
+def inject_event_list_converter(application_scope):
+    """Return an event list converter configured by an application scope."""
+    return EventListConverter(
+        inject_event_converter(application_scope),
+        inject_event_suppressor(application_scope))
+
+
 def inject_event_converter(application_scope):
     """Return an event converter configured by an application scope."""
     return EventConverter(
@@ -198,6 +206,20 @@ def inject_oauth2_session_starter(application_scope):
         application_scope.xibo_client_secret,
         inject_xibo_token_url(application_scope),
         inject_user_agent(application_scope))
+
+
+def inject_event_suppressor(application_scope):
+    """Return an event suppressor configured by an application scope."""
+    return application_scope.event_suppressor(
+        inject_event_suppressor_provider(application_scope))
+
+
+def inject_event_suppressor_provider(application_scope):
+    """Return a function that provides an event suppressor configured by an
+    application session scope."""
+    def get():
+        return EventSuppressor(application_scope.suppressed_event_ids)
+    return get
 
 
 def inject_enter_xibo_session_scope(application_scope):
@@ -306,6 +328,7 @@ def inject_event_updater_provider(application_scope, xibo_session_scope):
             xibo_events,
             xibo_event_crud,
             inject_anti_flapper(application_scope),
+            inject_event_suppressor(application_scope),
             inject_special_location_monitor(application_scope)
             )
     return get
@@ -393,9 +416,10 @@ def inject_meetup2xibo(application_scope):
     return Meetup2Xibo(
         inject_meetup_events_retriever(application_scope),
         inject_selected_conflict_analyzer(application_scope),
-        inject_event_converter(application_scope),
+        inject_event_list_converter(application_scope),
         inject_site_cert_assurer(application_scope),
         inject_oauth2_session_starter(application_scope),
+        inject_event_suppressor(application_scope),
         inject_enter_xibo_session_scope(application_scope),
         )